adds pull/build pattern
This commit adds a pattern that pulls an image before it gets build.
Thus even if the last image is not on the local drive it the build cache
can be levereaged because of the pull. Therefore a latest tag is always
created and pushed. This tag will be pulled.